Add bidder, commenter info and refine queries
Added 'bidder' field to the 'Bid' model and 'commenter', 'date_created' fields to the 'Comment' model to track who made a bid or a comment and when the comment was made. Changed the 'starting_bid' field in 'AuctionListing' model and 'amount' field in 'Bid' model from IntegerField to DecimalField to allow decimal amounts. Listings and bids are now filtered and sorted appropriately in the views. This updates enables more accurate tracking of user activity and enhances data accuracy for bids.