Update create_listing function for user context and forms.py for efficient code
Updated the create_listing function to directly create an AuctionListing object so as to include the creator of the listing directly from the request user. The object is created with cleaned data received form the auction form. Also removed 'active' field from the form and managed its value in views.py for better control over the initial status of auction. Finally, there are minor codebase refinements like removing unnecessary spaces and lines for cleaner code.