Refactor bid view and improve bid validation
This commit refactors the bid view in auctions' app for better readability and exception handling. Previously, all code was nested under 'if request.method == "POST":"' condition and the bid form's validation errors were not well-handled. The changes improve the view structure and introduce a more clear condition flow. The bid validation was also improved. Now, the code checks if the bid is higher not only than the previous highest bid but also than the listing's starting bid.