Improve comment system by utilizing AJAX
Updated the auction listing HTML to use AJAX for comment submission and loading. This ensures that page doesn't need to be refreshed each time a comment is submitted. Introduced a new route 'get-comments' to fetch comments data in JSON format, which is used by the AJAX call. Modified the 'comment' view to return JsonResponse after comment has been successfully saved, which is needed by the AJAX call to confirm the success of comment submission.