Improve layout and user interface of auction listings
In the layout file, unnecessary comments were removed and jQuery UI and its CSS file were added. This was done to improve the site's responsiveness and to implement collapsible auction listings for a better user experience and cleaner interface.

In the index file, auction listings are now shown as a collapsible, accordion-style list. This was done using jQuery UI's accordion feature for a more organized view for users. The scripts to implement this feature were also added at the end of the file.