Add category information and improve image presentation
A new function 'category_info' was added in 'context_processors.py' to provide info about whether the current request path starts with '/category/'. This data is then used to manipulate the template 'layout.html'.

In the 'uaa490styles.css' file, a new class '.detail-img' was added to handle the presentation of the images in a more detailed view, which is used in 'listing.html'.

Setting for 'category_info' was added in 'settings.py'.

The buttons in 'create_listing.html' were slightly adjusted for uniformity and redundant links removed.

All the changes aim to improve user navigation and user interface.